requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pyobjc-framework-Cocoa>=10.3
pyobjc-framework-Quartz>=10.3
pyobjc-framework-ApplicationServices>=10.3
